penambahan add tindakan

+++ src/com/neurix/simrs/transaksi/tindakanrawat/model/ItSimrsTindakanRawatEntity.java
@@ -1,6 +1,7 @@
 package com.neurix.simrs.transaksi.tindakanrawat.model;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 public class ItSimrsTindakanRawatEntity implements Serializable {
@@ -11,9 +12,9 @@ public class ItSimrsTindakanRawatEntity implements Serializable {
     private String namaTindakan;
     private String idDokter;
     private String idPerawat;
-    private Long tarif;
-    private Long qty;
-    private Long tarifTotal;
+    private BigInteger tarif;
+    private BigInteger qty;
+    private BigInteger tarifTotal;
     private String flag;
     private String action;
     private Timestamp createdDate;
@@ -69,27 +70,27 @@ public class ItSimrsTindakanRawatEntity implements Serializable {
         this.idPerawat = idPerawat;
     }
 
-    public Long getTarif() {
+    public BigInteger getTarif() {
         return tarif;
     }
 
-    public void setTarif(Long tarif) {
+    public void setTarif(BigInteger tarif) {
         this.tarif = tarif;
     }
 
-    public Long getQty() {
+    public BigInteger getQty() {
         return qty;
     }
 
-    public void setQty(Long qty) {
+    public void setQty(BigInteger qty) {
         this.qty = qty;
     }
 
-    public Long getTarifTotal() {
+    public BigInteger getTarifTotal() {
         return tarifTotal;
     }
 
-    public void setTarifTotal(Long tarifTotal) {
+    public void setTarifTotal(BigInteger tarifTotal) {
         this.tarifTotal = tarifTotal;
     }
